Add files via upload
yesterday I created this repository from linux system to save my basic html codes from the freecodecamp but i couldn"t pull it to my windows so i fork it and code and in the end i pushed the files back to main. it was successful but nothing shown in here so today i drag and drop the html file now. 24.11.2020.